requirements.txt
# fastapi==0.103.1
# uvicorn==0.23.2
# pydantic==2.3.0
# firebase-admin==6.2.0
# openai==1.3.0
# python-jose==3.3.0
# passlib==1.7.4
# python-multipart==0.0.6
# duckduckgo-search
# python-dotenv==1.0.0
# email-validator==2.0.0
# httpx<0.27
# pydantic_settings
# ipython
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
firebase-admin==6.4.0
openai==1.3.6
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
duckduckgo-search==3.9.6
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.7